backend/api/requirements.txt
# Dufour Middleware API - Python Dependencies

# Web Framework
fastapi==0.103.2
uvicorn[standard]==0.24.0
python-multipart==0.0.9
httpx==0.26.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
asyncpg==0.29.0  # Async PostgreSQL driver

# Testing
pytest==8.0.0
pytest-asyncio==0.23.5
pytest-mock==3.12.0

# Geospatial libraries
fiona==1.9.5
shapely==2.0.2
pyproj==3.6.1

# XML/JSON (for parsing QGIS project files)
lxml==5.1.0
pydantic==2.5.3

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1

# Image processing (for print composition with military symbols)
Pillow==10.2.0
